requirements.txt
Django==5.0.4
django-filter==24.2
django-mptt==0.16.0
Markdown==3.6
djangorestframework==3.15.1
